Move krate spans to separate module
